+++ src/main/java/com/project/cheerha/domain/bookmark/dto/request/DeleteBookmarkRequestDto.java
@@ -1,8 +1,8 @@
-package com.project.cheerha.domain.bookmark.dto;
+package com.project.cheerha.domain.bookmark.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
-public record CreateBookmarkRequestDto(
+public record DeleteBookmarkRequestDto(
         @NotNull
         Long jobOpeningId
 ) {
