+++ src/main/java/com/project/cheerha/domain/bookmark/dto/response/ReadBookmarkResponseDto.java
@@ -1,4 +1,4 @@
-package com.project.cheerha.domain.bookmark.dto;
+package com.project.cheerha.domain.bookmark.dto.response;
 
 import com.project.cheerha.domain.bookmark.entity.Bookmark;
 import com.project.cheerha.domain.jobOpening.entity.JobOpening;
